v1_parse_cleanup
