+++ tests/scripts/config_directory/raw/users_with_pk.preql
@@ -1,5 +1,5 @@
 # Datasource ingested from users_with_pk
-# Generated on 2026-04-14 20:23:17.787305
+# Generated on 2026-04-15 00:22:06.656646
 import std.net;
 
 key user_id int;
