+++ tests/test_parse_engine.py
@@ -1,14 +1,8 @@
 from pytest import raises
 
-from trilogy.core.exceptions import UndefinedConceptException
+from trilogy.core.exceptions import InvalidSyntaxException, UndefinedConceptException
 from trilogy.core.models.environment import Environment
-from trilogy.parsing.parse_engine import (
-    ERROR_CODES,
-    PARSER,
-    InvalidSyntaxException,
-    ParseToObjects,
-    unpack_visit_error,
-)
+from trilogy.parsing.parse_engine_v2 import ERROR_CODES
 
 TEXT = """
 const a <- 1;
@@ -22,34 +16,12 @@ select
 
 def test_parser():
     env = Environment()
-    x = ParseToObjects(environment=env)
-    x.set_text(TEXT)
-
-    failed = False
-    try:
-        tokens = PARSER.parse(TEXT)
-        x.transform(tokens)
-        x.run_second_parse_pass()
-    except Exception as e:
-        failed = True
-        with raises(UndefinedConceptException):
-            unpack_visit_error(e)
-    assert failed
-
-
-TEXT2 = """
-const a <- 1;
-
-select
-    a,
-FROM a
-;
-"""
+    with raises(UndefinedConceptException):
+        env.parse(TEXT)
 
 
 def test_parse_datatype_in_datasource():
     env = Environment()
-    x = ParseToObjects(environment=env)
     test_text = """
 key x int;
 property x.timestamp timestamp;
@@ -60,11 +32,7 @@ datasource funky (
 address fun;
 
 """
-    x.set_text(test_text)
-
-    tokens = PARSER.parse(test_text)
-    x.transform(tokens)
-    x.run_second_parse_pass()
+    env.parse(test_text)
 
 
 def test_from_error():
@@ -128,7 +96,7 @@ def test_semicolon_error():
 
     select
         a+2 as fun,
-    
+
     """
     with raises(InvalidSyntaxException) as e:
         env.parse(TEXT2)
