+++ trilogy/__init__.py
@@ -4,7 +4,7 @@ from trilogy.dialect.enums import Dialects
 from trilogy.executor import Executor
 from trilogy.parser import parse
 
-__version__ = "0.3.226"
+__version__ = "0.3.227"
 
 __all__ = [
     "parse",
