+++ trilogy/parsing/helpers.py
@@ -1,6 +1,15 @@
 from __future__ import annotations
 
+from typing import Protocol
+
 from trilogy.core.models.author import Comment
+
+
+class Meta(Protocol):
+    line: int | None
+    column: int | None
+    end_line: int | None
+    end_column: int | None
 
 
 def comment_body(comment: Comment) -> str:
