+++ trilogy/parsing/v2/rules/token_rules.py
@@ -179,6 +179,10 @@ def PARSE_F_FILE_PATH(token: SyntaxToken, context: RuleContext) -> str:
     return _interpolate_params(token.value[2:-1], context)
 
 
+def PARSE_F_QUOTED_ADDRESS(token: SyntaxToken, context: RuleContext) -> str:
+    return _interpolate_params(token.value[2:-1], context)
+
+
 def PARSE_CHART_TYPE(token: SyntaxToken, context: RuleContext) -> str:
     return token.value.lower()
 
@@ -232,6 +236,7 @@ TOKEN_HYDRATORS: dict[SyntaxTokenKind, TokenHydrator] = {
     SyntaxTokenKind.PERSIST_MODE: PARSE_PERSIST_MODE,
     SyntaxTokenKind.FILE_PATH: PARSE_FILE_PATH,
     SyntaxTokenKind.F_FILE_PATH: PARSE_F_FILE_PATH,
+    SyntaxTokenKind.F_QUOTED_ADDRESS: PARSE_F_QUOTED_ADDRESS,
     SyntaxTokenKind.CHART_TYPE: PARSE_CHART_TYPE,
     SyntaxTokenKind.VALIDATE_SCOPE: PARSE_VALIDATE_SCOPE,
     SyntaxTokenKind.COPY_TYPE: PARSE_COPY_TYPE,
